+++ lib/oneview-sdk/resource/api200/enclosure_group.rb
@@ -49,9 +49,11 @@ module OneviewSDK
 
       # Adds the logical interconnect group
       # @param [OneviewSDK::LogicalInterconnectGroup] lig Logical Interconnect Group
+      # @raise [OneviewSDK::NotFound] if the LIG uri is not set and cannot be retrieved
       # @return [OneviewSDK::EnclosureGroup] self
       def add_logical_interconnect_group(lig)
         lig.retrieve! unless lig['uri']
+        raise(NotFound, 'LIG not found!') unless lig['uri']
         lig['interconnectMapTemplate']['interconnectMapEntryTemplates'].each do |entry|
           entry['logicalLocation']['locationEntries'].each do |location|
             add_lig_to_bay(location['relativeValue'], lig) if location['type'] == 'Bay' && entry['permittedInterconnectTypeUri']